Fix tests and switch edge weights to floats

+++ source/core/include/blast/graph.hpp
@@ -8,7 +8,7 @@ namespace blast {
 	class Graph {
 		std::vector<std::shared_ptr<Node>> nodes;
 		// List of adjacency rows with pairs of weight and index of neighbor node
-		std::vector<std::vector<std::pair<int, size_t>>> adj_list;
+		std::vector<std::vector<std::pair<float, size_t>>> adj_list;
 	public:
 		Graph() = default;
 		Graph(const Graph& original);
@@ -16,11 +16,11 @@ namespace blast {
 		void add_node(const std::shared_ptr<Node>& node);
 		void remove_node(size_t index);
 		std::shared_ptr<Node> get_node(size_t index) const;
-		void add_directed_edge(size_t from, size_t to, int weight);
-		void add_undirected_edge(size_t a, size_t b, int weight);
+		void add_directed_edge(size_t from, size_t to, float weight);
+		void add_undirected_edge(size_t a, size_t b, float weight);
 		void remove_directed_edge(size_t from, size_t to);
 		void remove_undirected_edge(size_t a, size_t b);
-		[[nodiscard]] int get_edge_weight(size_t from, size_t to) const;
+		[[nodiscard]] float get_edge_weight(size_t from, size_t to) const;
 		[[nodiscard]] bool exists_edge(size_t from, size_t to) const;
 		[[nodiscard]] size_t get_node_count() const;
 		void clear();
@@ -28,5 +28,5 @@ namespace blast {
 	};
 
 	int get_length_of_path(const Graph& graph, const std::vector<size_t>& path, bool loop);
-	void add_edges_from_adjacency_matrix(Graph& graph, const std::vector<std::vector<int>>& adj_matrix);
+	void add_edges_from_adjacency_matrix(Graph& graph, const std::vector<std::vector<float>>& adj_matrix);
 }; // namespace blast
